Add files via upload
In this project, you will launch an **Amazon EC2 instance** using **Amazon Linux 2023** and configure it as a web server with Apache (httpd)**. You will create a customized webpage** with a visually appealing design. The setup will be automated using **User Data**, ensuring that the web server and webpage are configured automatically upon instance startup.